refactor(env): Extract `mdn` into separate file
